move Product related components into new  _molecules/ dir
